refactor: consolidate duplicate CacheKeyGeneratorService implementations
- Merged 205-line and 61-line implementations into single service
- Added semantic signature generation from old cache-key-generator-service.ts
- Enhanced text normalization with stopword removal
- Updated imports in ai-pattern-detector.ts and cache-warming-rule-service.ts
- Deleted duplicate file: lib/services/cache-key-generator-service.ts (61 lines)
- Created OptimizedCacheKeyOptions interface for type safety

Business Impact: Single source of truth, eliminated confusion, consistent caching

Files Modified:
- lib/services/cache/key-generator-service.ts (+41 lines, comprehensive merge)
- lib/services/ai-pattern-detector.ts (import updated)
- lib/services/cache-warming-rule-service.ts (import updated)
- lib/services/cache-key-generator-service.ts (deleted)

Architectural Improvement: Dependency cleanup reduces maintenance burden
